requirements.txt
playwright==1.49.1
pillow==11.1.0
numpy==2.2.3
beautifulsoup4==4.13.3
httpx==0.28.1
python-dotenv==1.0.1
fastapi==0.115.8
uvicorn[standard]==0.34.0
apscheduler==3.11.0
pydantic==2.10.6
pydantic-settings==2.7.1
